feat: Allow to login with username

+++ prisma/seed.ts
@@ -3,13 +3,18 @@ import prisma from "../lib/prisma";
 
 async function main() {
   try {
-    await prisma.user.deleteMany();
+    if (process.env.NODE_ENV !== "production") {
+      await prisma.user.deleteMany();
+
+      console.log("🧹 Deleted all data seeded successfully");
+    }
 
     await prisma.user.create({
       data: {
         name: "John",
         surname: "Doe",
         email: "jdoe@me.com",
+        username: "jdoe",
         password: await hash("jdoe", Number(process.env.SALT_ROUNDS)),
         isActivated: true,
       },
@@ -17,7 +22,7 @@ async function main() {
 
     prisma.$disconnect();
 
-    console.log("Seeded successfully");
+    console.log("✅ Seeded successfully");
   } catch (e: any) {
     prisma.$disconnect();
     console.error(e);
